app/requirements.txt
fastapi
uvicorn
psycopg2-binary
ollama
python-dotenv
asyncpg
aiosqlite
google-generativeai